Embed associated post type records in Taxonomy objects
This is the first pass at embedding a list of associated post types in
taxonomy records, as described in #197. Types are NOT embedded if they
would be added to a taxonomy record that is itself embedded.

@rmccue has proposed changing how the $_in_taxonomy behavior works to
use a "$context" variable, as in other methods; this commit should be
considered a work-in-progress until he has signed off on it.