Ensure post deletion is idempotent
Practically speaking: a DELETE request to a post resource will ensure
the post is trashed. If the post is already trashed, no change will
occur. To permanently delete, pass force=true, and it will be
permanently deleted regardless of status.

This is a change from the previous behaviour of the first DELETE
trashing the post, and the second DELETE permanently deleting it.