Rename the `wp_json_server_before_serve` to `wp_json_init`
Currently, developers must use the `wp_json_server_before_serve` hook to add any new endpoints. This is very verbose and difficult to remember, I think it would be better to have this named akin to `init` and `admin_init` hooks. In the future, if the plugin switches to `wp_api` prefix, rather than `json` I think we'd change this hook to `api_init` but for now I think it's best kept inline with the rest of the naming conventions.