Refine  permissions check for getting lists of collection items.
On advice of @rachelbaker and @danielbachhuber, refine the permissions test to require the user has the edit_posts permission for the specific post_type. Also add a unit test to confirm the functionality of allowing the post list to be retrieved when the user does in fact have the editing capability.