Power the Comments endpoint args from the schema, like everything else
The previous args had defaults of empty strings for `wp_allow_comment`,
however I see this as an incorrect use of the abstraction. Args should
not have defaults just to please an underlaying function that expects
those values.

This also needed to move into `create_item` as the same default can not
be used for the `update_item` endpoint args.