Add the ability to assign more than one term at once to a post
This is currently difficult to do, as you have to send one request for
each associtation. Instead, this PR makes it possible to POST to
`/wp/v2/posts/123/terms/category` with `{ ids: [1,2,4] }`.

We could also add a `names` param to assign multiple terms via Name,
which is an often requested feature we don't currently support.