Fix MD5 check on file uploads.
Upload from file checking for header 'CONTENT_MD5', but all headers are
canonicalized by WP_REST_Request::canonicalize_header_name(), so this
header is never present.

Also, upload from file tries to handle the header as a string instead of
as an array.

Test case for bad MD5 header incorrectly checks for 201, instead of 412.

This commits fixes the MD5 check on upload from file by copying the way it
is handled by upload from data, that correctly checks for 'content_md5'
header and use it as an array.