First round at turning protected methods into private methods
Our controller classes are designed to expose a consistent abstraction.
The expected public methods include `get_items()`, `get_item()`,
`create_item()`, `update_item()`, `delete_item()`,
`get_items_permissions_check()`, and so on.

Methods unique to a specific class should be marked `private` instead
of `protected`, so we don't expose them as interfaces for subclasses and
retain the freedom to refactor the internal behavior of a class at a
future date.