Adding validation
Coolest part was I found out that this year is a leap year from this
code and testing! Do
/wp/v2/posts/?date_query[0][year]=2015&date_query[0][month]=2&date_query[0][day]=29.
If it doesn't fail it's a leap year! Hooray.

Added validation.  This is still not error proof as I need to handle
weird cases with how the top level params work, like when monthnum is
specified without year.  Maybe chopping top level parameter support
would be nice, but would leave this incomplete so I will continue to add
error handling for those strange cases.