#2331 Fix links in plugin docs
This is a partial fix to #2331.  This fixes the link for the WP REST API
TEAM to point to v2 and fixes a link in the readme portion of the plugin
to point to the v2 plugin instead of v1.  Because there is a plugin
named wp-api in the WordPress.org plugin repository it is matching the
View Details link to that plugin which has zero relevance to WP REST API
v2.