#538-Add-unauthenticated-user-creation
For #538. The approach taken here allows for unauthenticated user
creation (i.e. a new user registering for a website). Things to
consider: how we want to handle multisite user creation. I think what we
have right now works well however, I am not sure how multi-site users
should be handled.  Should add_user_to_blog() be added somewhere and
have a blogs attribute added to the schema, containing a list of blog
id's a user is associated with? Also potentially add in validation for a
format password to the password arg so that only valid WordPress
passwords would be accepted. Thoughts in general would be appreciated.