Fix-#2194-validate-array-type
Fixes #2194.  However, there may be a better implementation but I
believe this is a quick solution and works for csv lists as well as long
as they are sanitized via wp_parse_id_list() or a function that returns
sanitized arrays.   Originally had a more complex system allowing for an
array format to be specified like 'format' => 'id_list', but it quickly
descended into madness.