Fixes #386 roles param for GET users
Fixes #386 roles param for get user.  Exposes role querying for GET
wp/v2/users endpoint.  This currently support an OR relation the default
is AND but I thought using the OR operator would match up with what
seems to be how people would want to use it.  AND didn't seem to make
much sense as out of the box WordPress doesn't support multiple roles
for a user if I am correct?