Fixing tests.
Thank you danielbachuber.  Test now works, but I am not sure if the test
should work differently.  Currently it only checks if a leading slash is
on the path which it should be.  However, it is difficult to get the
untouched data checked to see if it has a leading slash or not, which is
the intention of the test.