Fixes #1612 by adding link to collection sharing comment parent.
Fixes #1612,

This patch adds a `children` link to the response when the comment has
children.  The link is to a collection of comments who share the same
parent as the original comment resource.  This enables a client to then
recursively move through the links, with pagination parameters to boot,
to performantly collect all threaded comments of a single comment.