Alter default comment sort order to be "desc"
This follows discussion in #2588 and [in slack][slack-log] around
standardizing the sort order of the comments endpoint to bring it
in line with both the other endpoints in the API, and with
`WP_Comment_Query`, by switching back to "desc" as the default sort
order. Note that is a breaking change if API consumers are relying
on comment order to be ascending by default.

Closes #2588

[slack-log]: https://wordpress.slack.com/archives/core-restapi/p1473088762000227