Update and add unit tests to account for ignoring filter[posts_per_page]
@joehoyle @rmccue fixed the test broken in #2699 (the test presumes that
filter properties work consistently, which that PR breaks by ignoring
posts_per_page entirely), but I am not sure how to handle the test for
`filter[posts_per_page]=-1`: should I be testing that it is ignored, as
seems to be the case, or should I be testing that it is an error?

The solution in #2699 is the quick fix but seems to introduce confusion
around which `filter` properties are valid, and which aren't.