Add failing tests to demonstrate lack of auto-draft support
For #2737

The list of statuses supported by the API omits auto-draft, as it
is considered an internal status and it seems we are using
`get_post_stati` with `'internal' => false` to return our list of
valid API status objects. I am curious what the team thinks the
right thing to do here might be?