Fix error when a single meta value attempts to update to the current value. Adds test.
If the value to be updated is the same as is currently in the database, update_metadata() returns false which leads to the error. The fix is to have the exact same check run in update_meta_value and return true if the values are the same. It doesn't update. Also included is a test for adding a duplicate value.