Properly register "{taxonomy}_exclude" parameters in the schema
PR #2756 added an `_exclude` complement to every registered taxonomy
query parameter method, e.g. `categories_exclude`, to permit an
API consumer to list an array of term IDs with which returned posts
should NOT be associated. However, that PR omitted the important step
of registering the parameters on the schema! This PR does so.