Return empty taxonomy collections as JSON object, not array
The response from `/wp/v2/taxonomies` is a JSON object, keyed with the
name of each taxonomy. If you use the `?type=` query parameter to specify
an object for which no taxonomies are available, or to specify an invalid
type of object, the empty PHP `array()` will be rendered in JSON as `[]`.

To ensure that the type of this resource's response remains consistent
this pull request casts empty arrays to a PHP object, which will be
correctly rendered as `{}` in the response JSON.

Fixes #2803