Send an error if the client tries to delete an option with an non-scaler value.
This is to cover the case when there is an invalid option value in the
database that does not match the one in register_setting().

Currently we are only doing this check for setting values to `null` to
work around the Client-sends-the-same-response issue, but we _do_ all
setting the value to some other value.